Rename & move the "Tat" structure somewhere I can document it
Now named InMemoryState, it seems more appropriate.
