+++ src/state/in_memory.rs
@@ -1,0 +1,59 @@
+use crate::lib::*;
+
+use crate::nanos::Nanos;
+use crate::state::{NotKeyed, StateStore};
+
+/// An in-memory representation of a GCRA's rate-limiting state.
+///
+/// Implemented using [`AtomicU64`] operations, this state representation can be used to
+/// construct rate limiting states for other in-memory states: e.g., this crate uses
+/// `InMemoryState` as the states it tracks in the keyed rate limiters it implements.
+///
+/// Internally, the number tracked here is the theoretical arrival time (a GCRA term) in number of
+/// nanoseconds since the rate limiter was created.
+#[derive(Default)]
+pub struct InMemoryState(AtomicU64);
+
+impl InMemoryState {
+    pub(crate) fn measure_and_replace_one<T, F, E>(&self, f: F) -> Result<T, E>
+    where
+        F: Fn(Option<Nanos>) -> Result<(T, Nanos), E>,
+    {
+        let mut prev = self.0.load(Ordering::Acquire);
+        let mut decision = f(NonZeroU64::new(prev).map(|n| n.get().into()));
+        while let Ok((result, new_data)) = decision {
+            match self.0.compare_exchange_weak(
+                prev,
+                new_data.into(),
+                Ordering::Release,
+                Ordering::Relaxed,
+            ) {
+                Ok(_) => return Ok(result),
+                Err(next_prev) => prev = next_prev,
+            }
+            decision = f(NonZeroU64::new(prev).map(|n| n.get().into()));
+        }
+        // This map shouldn't be needed, as we only get here in the error case, but the compiler
+        // can't see it.
+        decision.map(|(result, _)| result)
+    }
+}
+
+/// The InMemoryState is the canonical "direct" state store.
+impl StateStore for InMemoryState {
+    type Key = NotKeyed;
+
+    fn measure_and_replace<T, F, E>(&self, _key: &Self::Key, f: F) -> Result<T, E>
+    where
+        F: Fn(Option<Nanos>) -> Result<(T, Nanos), E>,
+    {
+        self.measure_and_replace_one(f)
+    }
+}
+
+impl Debug for InMemoryState {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
+        let d = Duration::from_nanos(self.0.load(Ordering::Relaxed));
+        write!(f, "InMemoryState({:?})", d)
+    }
+}
